requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
Django==4.2.14
django-cors-headers==4.3.1
django-crontab==0.7.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
inflection==0.5.1
jiter==0.8.0
openai==1.56.2
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.8.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==1.0.0
ua-parser-builtins==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
user-agents==2.2.0
